# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documents",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sell_orders/create": {
            "post": {
                "tags": [
                    "Sell Orders"
                ],
                "summary": "Create Sell Orders",
                "operationId": "b4f2a438a9ffefac28ec981511cfc8a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "token_type",
                                    "amount",
                                    "per_price"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "string",
                                        "example": "abc527bd-79d0-4799-a883-9c7257990d3b"
                                    },
                                    "token_type": {
                                        "type": "string",
                                        "example": "BTC,ETH,XRP,DOGE"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "per_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1999.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "9390cea0-6fa6-41f3-a003-b5c3fb69db3a"
                                                },
                                                "create_date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-05-16 11:35:59"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Create sale order failed"
                    }
                }
            }
        },
        "/api/sell_orders/list": {
            "post": {
                "tags": [
                    "Sell Orders"
                ],
                "summary": "Sell Orders List",
                "operationId": "a5bf6a976f1e6675cc3c1a0f0fc4787c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "keyword": {
                                        "type": "string",
                                        "example": "Search by id"
                                    },
                                    "sort_by": {
                                        "type": "string",
                                        "example": "per_price,update_date"
                                    },
                                    "sort": {
                                        "type": "string",
                                        "example": "asc,desc"
                                    },
                                    "page": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "token_type": {
                                        "type": "string",
                                        "example": "BTC,ETH,XRP,DOGE"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sell_orders_id": {
                                                        "type": "string",
                                                        "example": "5ca2d56d-79ad-41e3-889f-0cdc12c77014"
                                                    },
                                                    "seller_id": {
                                                        "type": "string",
                                                        "example": "9390cea0-6fa6-41f3-a003-b5c3fb69db3a"
                                                    },
                                                    "token_type": {
                                                        "type": "string",
                                                        "example": "XRP"
                                                    },
                                                    "amount": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "per_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1999.99
                                                    },
                                                    "create_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-05-16 10:23:53"
                                                    },
                                                    "update_date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-05-16 10:23:53"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Not found data"
                    }
                }
            }
        },
...
    },
...
}